+++ src/soc/amd/stoneyridge/smihandler.c
@@ -16,6 +16,7 @@
 #include <amdblocks/smi.h>
 #include <amdblocks/smm.h>
 #include <elog.h>
+#include <types.h>
 
 /* bits in smm_io_trap   */
 #define SMM_IO_TRAP_PORT_OFFSET		16
@@ -31,7 +32,7 @@ static inline u16 get_io_address(u32 info)
 
 static void *find_save_state(int cmd)
 {
-	int core;
+	unsigned int core;
 	amd64_smm_state_save_area_t *state;
 	u32 smm_io_trap;
 	u8 reg_al;
@@ -220,7 +221,7 @@ static const struct smi_sources_t smi_sources[] = {
 
 void *get_smi_source_handler(int source)
 {
-	int i;
+	size_t i;
 
 	for (i = 0 ; i < ARRAY_SIZE(smi_sources) ; i++)
 		if (smi_sources[i].type == source)
